Share a few sample models and benchmarks.
The benchmarks cannot run all at the same time, and are controlled by a number of env variables, until we find a way to get all the platforms involved to work together.

This code was used to generate the poster graphs, and ran in GCloud instances.

Remaining TODO: Add instructions for running / installing dependencies.

PiperOrigin-RevId: 178765719