Desugaring list iter (#67)
Desugars `for x in a` loops

Detects if a for loop is being used that iterates directly over values of a variable (e.g. for a in x: res += a). Tangent needs to know that loop is implicitly indexing into x in order to propagate adjoint updates through to x, so we make the index access explicit.

Fixes #66